port/platform/common/automation/requirements.txt
PyYAML
invoke
psutil
requests
west
pyserial
pylink-square
coloredlogs
verboselogs
portalocker
lxml
dataclasses